requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
build==1.3.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
chromadb==1.1.1
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
durationpy==0.10
ecdsa==0.19.1
fastapi==0.118.2
filelock==3.20.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.2
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.78
langchain-deepinfra==0.1.0
langchain-openai==0.3.35
langchain-postgres==0.0.15
langchain-text-splitters==0.3.11
langsmith==0.4.33
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.23.1
openai==2.2.0
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pgvector==0.3.6
pillow==11.3.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.32.1
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg-pool==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.1.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websocket-client==1.9.0
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
